rail-recommender/recommender-service/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
pandas==2.0.3
numpy==1.24.4
scikit-surprise==1.1.4
python-dotenv==0.19.0
requests==2.26.0